Fix typing error and misplaced sentence
The sentence "This list is far from comprehensive, and is just meant to highlight a few of the different kinds of Transformer models. Broadly, they can be grouped into three categories:" was mistakenly added on the line where GPT-3 was introduced. Moved it after the timeline and before the three categories are specified.